+++ stories/components/select/select.stories.ts
@@ -3,6 +3,7 @@ import { html } from 'lit-html';
 // @ts-ignore
 import selectDocumentation from './select-documentation.md';
 import { getComponentArgs } from '../../utils/args-generation';
+import {WcsSelectSize} from "../../../src";
 
 const meta: Meta = {
     title: 'Components/Select',
@@ -18,13 +19,13 @@ const meta: Meta = {
 };
 export default meta;
 
-const Template: StoryFn<Partial<{ value: string, placeholder: string, disabled: boolean, multiple: boolean, chips: boolean }>> = (args) => html`
+const Template: StoryFn<Partial<{ value: string, placeholder: string, disabled: boolean, multiple: boolean, chips: boolean, size: WcsSelectSize }>> = (args) => html`
     <style>
         wcs-select{
             width: 400px;
         }
     </style>
-    <wcs-select placeholder="${args.placeholder}" id="leselect" value="${args.value}" ?disabled="${args.disabled}" ?multiple="${args.multiple}" ?chips="${args.chips}">
+    <wcs-select placeholder="${args.placeholder}" id="leselect" value="${args.value}" ?disabled="${args.disabled}" ?multiple="${args.multiple}" ?chips="${args.chips}" size="${args.size ?? 'm'}">
         <wcs-select-option value="1" chip-background-color="var(--wcs-pink)">One</wcs-select-option>
         <wcs-select-option value="2" chip-background-color="var(--wcs-yellow)" chip-color="var(--wcs-black)">Two</wcs-select-option>
         <wcs-select-option value="3" chip-background-color="var(--wcs-red)">Three</wcs-select-option>
@@ -37,8 +38,25 @@ Default.args = {
     placeholder: 'Le select par défaut',
     disabled: false,
     multiple: false,
-    chips: false
+    chips: false,
+    size: 'm'
 };
+
+export const Sizes: StoryFn<Partial<{ }>> = () => html`
+    <div style="display: flex; gap: var(--wcs-base-margin)">
+        <wcs-select placeholder="Select L" id="leselect-l" size="l" style="width: 200px;">
+            <wcs-select-option value="1">One</wcs-select-option>
+            <wcs-select-option value="2">Two</wcs-select-option>
+            <wcs-select-option value="3">Three</wcs-select-option>
+        </wcs-select>
+        <wcs-select placeholder="Select M (default)" id="leselect-m" size="m" style="width: 200px;">
+            <wcs-select-option value="1">One</wcs-select-option>
+            <wcs-select-option value="2">Two</wcs-select-option>
+            <wcs-select-option value="3">Three</wcs-select-option>
+        </wcs-select>
+    </div>
+`;
+Sizes.args = {};
 
 export const Disabled = Template.bind({});
 Disabled.args = {
@@ -46,7 +64,8 @@ Disabled.args = {
     placeholder: 'Le select désactivé',
     disabled: true,
     multiple: false,
-    chips: false
+    chips: false,
+    size: 'm'
 };
 
 export const OneOptionSelected = Template.bind({});
@@ -55,7 +74,8 @@ OneOptionSelected.args = {
     placeholder: 'Le select avec une option selectionnée',
     disabled: false,
     multiple: false,
-    chips: false
+    chips: false,
+    size: 'm'
 };
 
 export const MultipleMode = Template.bind({});
@@ -64,7 +84,8 @@ MultipleMode.args = {
     placeholder: 'Le select en mode multiple',
     disabled: false,
     multiple: true,
-    chips: false
+    chips: false,
+    size: 'm'
 };
 
 export const MultipleAndChipsMode = Template.bind({});
@@ -73,7 +94,8 @@ MultipleAndChipsMode.args = {
     placeholder: 'Le select en mode multiple',
     disabled: false,
     multiple: true,
-    chips: true
+    chips: true,
+    size: 'm'
 };
 
 const oneOptionDisabledTemplate: StoryFn<Partial<{ value: string, placeholder: string, disabled: boolean, multiple: boolean, chips: boolean}>> = (args) => html`