+++ CHANGELOG.md
@@ -12,6 +12,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - **button**: add `target` (`"_blank" | "_self"`, null by default) to choose where to open links when using `href`
 - **button**: add the following sizes : `l`, `m (default)`, `s`
 - **input**: add the following sizes : `l`, `m (default)`, `s (for grids)` 
+- **select**: add the following sizes : `l`, `m (default)`
 - **doc**: add documentation on how to initialize an application with the business or communication design systems
 - **ci**: add the ability to deploy any branch on a dedicated environment
 - **textarea**: add the property `resize` with the folowing options : `none`, `both`, `horizontal`, `vertical`